requirements.txt
asgiref==3.3.1
Django==3.1.4
django-environ==0.4.5
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-dotenv==0.15.0
pytz==2020.5
sqlparse==0.4.1
whitenoise==5.2.0
gunicorn==20.0.4
pagseguro==0.3.4